Retrieving profile information is refactored and belongs to the controller now.
