+++ server/src/api/controllers/Auth.controller.js
@@ -1,12 +1,16 @@
 import models from "../models/index";
+import response from "../../utils/sendResponse";
+import fetch from "node-fetch";
 import jwt from "jsonwebtoken";
 
 class AuthController {
   constructor(model) {
+    this.BASE_URL = `https://discordapp.com/api`;
     this.model = model;
     this.oneWeek = 7 * 24 * 3600 * 1000;
 
     this.signIn = this.signIn.bind(this);
+    this.getProfile = this.getProfile.bind(this);
   }
 
   async findOrCreate({ id, username, avatar: _avatar }) {
@@ -45,6 +49,28 @@ class AuthController {
     res.redirect(`${process.env.BASE_PATH}?accessToken=${accessToken}`);
   }
 
+  async getProfile(req, res) {
+    const decoded = this.decodeToken(req);
+
+    if (!decoded) {
+      response(res, 404, {}, false);
+    }
+
+    const _response = await fetch(`${this.BASE_URL}/users/@me`, {
+      headers: {
+        Authorization: `Bearer ${decoded.accessToken}`,
+      },
+    });
+
+    const data = await _response.json();
+
+    response(res, 200, {
+      id: data.id,
+      username: data.username,
+      avatar: this.makeAvatarURL(data.id, data.avatar),
+    });
+  }
+
   makeAvatarURL(id, avatar) {
     return `https://cdn.discordapp.com/avatars/${id}/${avatar}.jpg`;
   }
@@ -68,6 +94,12 @@ class AuthController {
       secure: false, // https
       httpOnly: true,
     });
+  }
+
+  decodeToken(req) {
+    const _accessToken = req.headers["authorization"];
+    const accessToken = _accessToken.slice(7, _accessToken.length);
+    return jwt.verify(accessToken, process.env.ACCESS_TOKEN_SECRET);
   }
 }
